+++ mobile/cmp/input/Select.ts
@@ -272,8 +272,8 @@ class SelectInputModel extends HoistInputModel {
                 ? reactAsyncCreatableSelect
                 : reactAsyncSelect
             : creatableMode
-              ? reactCreatableSelect
-              : reactSelect;
+            ? reactCreatableSelect
+            : reactSelect;
     }
 
     @action